% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjcolors}
\alias{adjcolors}
\title{Convert a color with arbitrary transparency to a fixed color}
\usage{
adjcolors(colors, alpha)
}
\arguments{
\item{colors}{Color vectors.}

\item{alpha}{Alpha level ranging from 0 to 1.}
}
\value{
The colors with the specified alpha level.
}
\description{
This function takes a vector of colors and an alpha level and converts the colors
to fixed colors with the specified alpha level.
}
\keyword{internal}
