% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{adjust_network_layout}
\alias{adjust_network_layout}
\title{Adjust_network_layout}
\usage{
adjust_network_layout(
  graph,
  layout,
  width,
  height = 2,
  scale = 100,
  iter = 100
)
}
\description{
Adjust_network_layout
}
\keyword{internal}
