% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bg_layer}
\alias{bg_layer}
\title{Get a ggplot layer for background}
\usage{
bg_layer(
  data,
  x,
  palette,
  palcolor,
  alpha,
  keep_empty,
  facet_by,
  direction = "vertical"
)
}
\arguments{
\item{data}{A data frame}

\item{x}{A character string specifying the column name of the data frame to plot for the x-axis}

\item{palette}{A character string specifying the palette to use}

\item{palcolor}{A character string specifying the color to use in the palette}

\item{alpha}{A numeric value specifying the transparency of the plot}

\item{keep_empty}{A logical value indicating whether to keep empty groups}

\item{facet_by}{A character string specifying the column name(s) of the data frame to facet the plot}

\item{direction}{A character string specifying the direction for the background}
}
\value{
A ggplot layer for background
}
\description{
Get a ggplot layer for background
}
\keyword{internal}
