% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blend_colors}
\alias{blend_colors}
\title{Blend colors}
\usage{
blend_colors(colors, mode = c("blend", "average", "screen", "multiply"))
}
\arguments{
\item{colors}{Color vectors.}

\item{mode}{Blend mode. One of "blend", "average", "screen", or "multiply".}
}
\value{
The blended color.
}
\description{
This function blends a list of colors using the specified blend mode.
}
\keyword{internal}
