% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_just}
\alias{calc_just}
\title{Calculate hjust and vjust based on angle}
\usage{
calc_just(angle)
}
\arguments{
\item{angle}{A numeric value of the angle}
}
\value{
A list with h and v values
}
\description{
Calculate hjust and vjust based on angle
}
\keyword{internal}
