% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check the columns if columns found in the data}
\usage{
check_columns(
  df,
  columns,
  force_factor = FALSE,
  allow_multi = FALSE,
  concat_multi = FALSE,
  concat_sep = "_"
)
}
\arguments{
\item{df}{A data frame}

\item{columns}{A character vector of column names}

\item{force_factor}{Whether to force the columns to be factors}

\item{allow_multi}{Whether to allow multiple columns}

\item{concat_multi}{Whether to concatenate multiple columns}

\item{concat_sep}{The separator to use for concatenation}
}
\value{
A character string of the valid column
}
\description{
Check the columns if columns found in the data
}
\keyword{internal}
