% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_keep_empty}
\alias{check_keep_empty}
\title{check_keep_empty
Check and normalize keep_empty parameter}
\usage{
check_keep_empty(keep_empty, cols = NA)
}
\arguments{
\item{keep_empty}{keep_empty}

\item{cols}{column names if keep_empty is a single value}
}
\value{
normalized keep_empty
}
\description{
check_keep_empty
Check and normalize keep_empty parameter
}
\keyword{internal}
