% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_keep_na}
\alias{check_keep_na}
\title{check_keep_na
Check and normalize keep_na parameter}
\usage{
check_keep_na(keep_na, cols = NA)
}
\arguments{
\item{keep_na}{keep_na}

\item{cols}{column names if keep_na is a single value}
}
\value{
normalized keep_na
}
\description{
check_keep_na
Check and normalize keep_na parameter
}
\keyword{internal}
