% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_legend}
\alias{check_legend}
\title{check_legend
Check if the legend.position and legend.direction are valid}
\usage{
check_legend(
  legend,
  datas_name,
  which = c("legend.position", "legend.direction")
)
}
\arguments{
\item{legend}{The value legend.position or legend.direction}

\item{datas_name}{names of the split data}
}
\value{
named list containing legend names
}
\description{
check_legend
Check if the legend.position and legend.direction are valid
}
\keyword{internal}
