% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_palcolor}
\alias{check_palcolor}
\title{check_palcolor
Check if the palcolor can be properly used}
\usage{
check_palcolor(palcolor, datas_name)
}
\arguments{
\item{palcolor}{palcolor}

\item{datas_name}{name of the split data}
}
\value{
named list containing color names
}
\description{
check_palcolor
Check if the palcolor can be properly used
}
\keyword{internal}
