% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_palette}
\alias{check_palette}
\title{check_palette
Check if the palette can be properly used}
\usage{
check_palette(palette, datas_name)
}
\arguments{
\item{palette}{palette}

\item{datas_name}{names of the split data}
}
\value{
named list containing palette names
}
\description{
check_palette
Check if the palette can be properly used
}
\keyword{internal}
