% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine plots into one}
\usage{
combine_plots(
  plots,
  combine = TRUE,
  split_by = NULL,
  nrow = NULL,
  ncol = NULL,
  byrow = NULL,
  axes = NULL,
  axis_titles = NULL,
  guides = NULL,
  design = NULL,
  recalc_size = TRUE
)
}
\arguments{
\item{plots}{A list of plots}

\item{combine}{Whether to combine the plots into one}

\item{split_by}{The column name to split the plots by. When provided,
the combined data from all sub-plots is available via \code{p$data}.}

\item{nrow}{The number of rows in the combined plot}

\item{ncol}{The number of columns in the combined plot}

\item{byrow}{Whether to fill the plots by row}

\item{recalc_size}{Whether to re-calculate the size of the combined plot}
}
\value{
The faceted plot. If guess_size is TRUE, attr(p, "height") and attr(p, "width") will be set
}
\description{
Combine plots into one
}
\keyword{internal}
