% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datam1_terra}
\alias{datam1_terra}
\title{datam1_terra.rda: source terra data with dim 10148 * 33
has lon lat for maps example}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10148 rows and 33 columns.}
\usage{
datam1_terra
}
\description{
datam1_terra.rda: source terra data with dim 10148 * 33
has lon lat for maps example
}
\keyword{satellite}
