% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{.apply_orderby}
\alias{.apply_orderby}
\title{Apply ordering to a data frame based on a specified expression}
\usage{
.apply_orderby(df, orderby, by, sby, sby_levels)
}
\arguments{
\item{df}{A data frame to be ordered.}

\item{orderby}{A character string representing the expression to order by.}

\item{by}{A character string representing the column name to group by.}

\item{sby}{A character string representing the column name to group by for secondary ordering.}

\item{sby_levels}{A character vector representing the levels of the secondary grouping column.}
}
\value{
A data frame ordered based on the specified expression.
}
\description{
Apply ordering to a data frame based on a specified expression
}
\keyword{internal}
