% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedheatmap.R
\name{.build_linked_matrix}
\alias{.build_linked_matrix}
\title{Build a matrix from long-format data}
\usage{
.build_linked_matrix(
  data,
  rows_by,
  columns_by,
  values_by,
  cell_agg = NULL,
  values_fill = NA
)
}
\arguments{
\item{data}{A data frame}

\item{rows_by}{Column name for rows}

\item{columns_by}{Column name for columns}

\item{values_by}{Column name for values}

\item{cell_agg}{Aggregation function (default: nanmean)}

\item{values_fill}{Fill value for missing combinations}
}
\value{
A list with \code{matrix}, \code{row_names}, \code{col_names}
}
\description{
Pivots data from long to wide format, grouping by rows_by/columns_by
and aggregating values_by.
}
\keyword{internal}
