% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{.check_annotation}
\alias{.check_annotation}
\title{Check the annotation specification
The annotations should not be any of by or split_by, and the columns must exist in the data}
\usage{
.check_annotation(data, annotation, by, split_by, which)
}
\arguments{
\item{data}{The data for Heatmap plot}

\item{annotation}{The annotation mappings}

\item{by}{The column for rows or columns}

\item{split_by}{The column for row or column splitting}

\item{which}{If the annotation is for row or column}
}
\description{
Check the annotation specification
The annotations should not be any of by or split_by, and the columns must exist in the data
}
\keyword{internal}
