% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{.prepare_extent}
\alias{.prepare_extent}
\title{Prepare the extent for spatial plots}
\usage{
.prepare_extent(ext)
}
\arguments{
\item{ext}{A numeric vector of length 4 specifying the extent as \code{c(xmin, xmax, ymin, ymax)},
or a \code{SpatExtent} object from the \code{terra} package.}
}
\value{
A \code{SpatExtent} object if \code{ext} is a numeric vector, or the original \code{SpatExtent} if it is already one.
NULL is returned if \code{ext} is NULL.
}
\description{
Prepare the extent for spatial plots
}
\keyword{internal}
