% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{.reorder_anno_side}
\alias{.reorder_anno_side}
\title{Reorder annotation list so split annotation (split_by) is farthest from heatmap body
and name annotation (by) is closest. User annotations stay in between.
For all annotation sides, ComplexHeatmap renders first-element = farthest,
last-element = closest to the heatmap body.}
\usage{
.reorder_anno_side(x, by, split_by, side)
}
\arguments{
\item{x}{A list of annotations}

\item{by}{The name of the annotation used for row/column names}

\item{split_by}{The name of the annotation used for splitting rows/columns}

\item{side}{The annotation side ("top", "bottom", "left", "right")}
}
\value{
A reordered list of annotations
}
\description{
Reorder annotation list so split annotation (split_by) is farthest from heatmap body
and name annotation (by) is closest. User annotations stay in between.
For all annotation sides, ComplexHeatmap renders first-element = farthest,
last-element = closest to the heatmap body.
}
\keyword{internal}
