% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{.resolve_anno_aliases}
\alias{.resolve_anno_aliases}
\title{Resolve annotation aliases:
.row         -> rows_by
.rows.split  -> rows_split_by
.col/.column -> columns_by
.col.split/.column.split -> columns_split_by}
\usage{
.resolve_anno_aliases(lst, row_key, rsplit_key, col_key, csplit_key)
}
\arguments{
\item{lst}{A list of annotations, which may contain aliases for rows_by, rows_split_by, columns_by, and columns_split_by.}

\item{row_key}{The actual key for rows_by in the list.}

\item{rsplit_key}{The actual key for rows_split_by in the list.}

\item{col_key}{The actual key for columns_by in the list.}

\item{csplit_key}{The actual key for columns_split_by in the list.}
}
\value{
A list of annotations with aliases resolved to the actual keys.
}
\description{
Resolve annotation aliases:
.row         -> rows_by
.rows.split  -> rows_split_by
.col/.column -> columns_by
.col.split/.column.split -> columns_split_by
}
\keyword{internal}
