% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{enrich_example}
\alias{enrich_example}
\title{An example of clusterProfiler enrichment result}
\description{
An example of clusterProfiler enrichment result
}
\examples{
\dontrun{
if (interactive()) {
  data(geneList, package="DOSE")
  de <- names(geneList)[abs(geneList) > 1.5]
  enrich_example <- clusterProfiler::enrichPathway(gene=de, pvalueCutoff = 0.05, readable=TRUE)
  enrich_example <- as.data.frame(enrich_example)
}
}
}
\concept{data}
