% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{enrich_multidb_example}
\alias{enrich_multidb_example}
\title{A example of clusterProfiler enrichment result with multiple databases}
\description{
A example of clusterProfiler enrichment result with multiple databases
}
\examples{
\dontrun{
if (interactive()) {
  data(enrich_example, package="plotthis")
  enrich_example$Database <- "DB1"
  enrich_example2 <- enrich_example
  enrich_example2$Database <- "DB2"
  enrich_example2$ID <- paste0(enrich_example2$ID, "_DB2")
  enrich_example2$Description <- paste0(enrich_example2$Description, " (DB2)")
  enrich_multidb_example <- rbind(enrich_example, enrich_example2)
  # usethis::use_data(enrich_multidb_example)
}
}
}
\concept{data}
