% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{facet_plot}
\alias{facet_plot}
\title{Facetting a plot}
\usage{
facet_plot(plot, facet_by, facet_scales, nrow, ncol, byrow, recalc_size = TRUE)
}
\arguments{
\item{plot}{The plot to facet or a list list(plot, height, width) if guess_size is TRUE}

\item{facet_by}{The column(s) to split data by and plot separately or facet by
If NULL, no faceting will be done}

\item{facet_scales}{Whether to scale the axes of facets.}

\item{nrow}{The number of rows in facet_wrap}

\item{ncol}{The number of columns in facet_wrap}

\item{byrow}{Whether to fill the plots by row}

\item{recalc_size}{Whether to re-calculate the size of the plot}
}
\value{
The faceted plot. If guess_size is TRUE, attr(p, "height") and attr(p, "width") will be set
}
\description{
Facetting a plot
}
\keyword{internal}
