% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{gggrob}
\alias{gggrob}
\title{Get the grid.draw-able ggplot grob
The output from ggplotGrob can not be directly used in grid.draw, the position can not be set.
This function extracts the gTree from the ggplot grob.}
\usage{
gggrob(p, void = TRUE, nolegend = TRUE)
}
\arguments{
\item{p}{A ggplot object}

\item{void}{If TRUE, the theme_void will be added to the ggplot object}

\item{nolegend}{If TRUE, the legend will be removed from the ggplot object}
}
\value{
A gTree object
}
\description{
Get the grid.draw-able ggplot grob
The output from ggplotGrob can not be directly used in grid.draw, the position can not be set.
This function extracts the gTree from the ggplot grob.
}
\keyword{internal}
