% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthis.R
\name{ggmap.plot.nemia}
\alias{ggmap.plot.nemia}
\title{plot.nemia.map. plot NEMIA region (NY state) using ggmap (google map).}
\usage{
ggmap.plot.nemia(data0)
}
\description{
plot.nemia.map. plot NEMIA region (NY state) using ggmap (google map).
}
