% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{ggmap.plot.us}
\alias{ggmap.plot.us}
\title{us map using ggmap}
\usage{
ggmap.plot.us(data0)
}
\arguments{
\item{data0}{data with lon and lat}
}
\description{
us map using ggmap
}
\examples{
ggmap.plot.us(datam1_terra)
}
