% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthis.R
\name{ggmap.plot.us.size}
\alias{ggmap.plot.us.size}
\title{The US map with fill = size0.}
\usage{
ggmap.plot.us.size(data0 = aeronetdt, size0 = "Ndays", sat0 = sat,
  NEMIA = F)
}
\description{
The US map with fill = size0.
}
