% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper.dilute.data}
\alias{helper.dilute.data}
\title{dilute dataset.}
\usage{
helper.dilute.data(data0, dilute_by = 10)
}
\arguments{
\item{data0}{dataset}

\item{dilute_by}{default to 10, 1/10}
}
\description{
help function
}
