% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper.pack.bins}
\alias{helper.pack.bins}
\title{Greedily pack objects with sizes given by \code{object.sizes}
into \code{n.bins} bins. Return a vector of the bin indices.
by Kodi Arfer.}
\usage{
helper.pack.bins(object.sizes, n.bins)
}
\description{
Greedily pack objects with sizes given by \code{object.sizes}
into \code{n.bins} bins. Return a vector of the bin indices.
by Kodi Arfer.
}
