% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{import_font}
\alias{import_font}
\title{Import a font}
\usage{
import_font(font)
}
\arguments{
\item{font}{Font family name to import.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Imports a font for use in plots.
This function imports a font from the \code{plotthis} package,
system fonts, or Google fonts.
This is borrowed from the \code{plummy} package.
To check available fonts, use \code{list_fonts()}.
}
