% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{join_heatmap_meta}
\alias{join_heatmap_meta}
\title{Join the meta data to the main data frame for heatmap}
\usage{
join_heatmap_meta(data, meta_data, by, cr_split_by, split_by, which)
}
\arguments{
\item{data}{A data frame containing the main data for the heatmap.}

\item{meta_data}{A data frame containing the meta data to be joined.}

\item{by}{A character string specifying the column name in \code{meta_data} to join on.
Either \code{rows_by} or \code{columns_by} should be specified in \code{data}.}

\item{cr_split_by}{A character string specifying the column name in \code{data} to join on.
Either \code{rows_split_by} or \code{columns_split_by} should be specified in \code{data}.}

\item{split_by}{A character string specifying the column name in \code{data} to join on.
Used to split the data into multiple heatmaps.}

\item{which}{A character string specifying whether to join on rows or columns.
Can be either \code{"row"} or \code{"column"}.}
}
\value{
A data frame with the meta data joined to the main data.
}
\description{
Join the meta data to the main data frame for heatmap
}
\keyword{internal}
