% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{list_fonts}
\alias{list_fonts}
\title{List fonts}
\usage{
list_fonts(
  family_only = FALSE,
  source = c("all", "builtin", "system", "google")
)
}
\arguments{
\item{family_only}{Only list the font family names.}

\item{source}{Source of the fonts to include. Can be one of "builtin", "system", "google", or "all".
Default is "all".}
}
\value{
A list of font family names, where the names are the source of the fonts
(system or google) if \code{family_only} is \code{TRUE}. If \code{family_only} is \code{FALSE},
returns a data frame with font family names and their paths.#'
}
\description{
Lists available fonts.
This function lists the fonts available in the \code{plotthis} package,
as well as system fonts and Google fonts if specified.
}
\details{
This is inspired from the \code{plummy} package.
Those fonts will be automatically imported when used in \code{theme_this()}, typically via the \code{theme_args}
argument in plotting functions. For example, \code{BarPlot(..., theme_args = list(font_family = "Barlow"))}.
You can also use \code{import_font("Barlow")} to import the font manually before using it in plots separately.
For example, \code{BarPlot(...) + theme(text = element_text(family = "Barlow"))}.
See \code{\link{import_font}()} for more details on importing fonts.
}
\examples{
\donttest{
list_fonts(family_only = TRUE, source = "all")
list_fonts(family_only = FALSE, source = "builtin")
list_fonts(family_only = FALSE, source = "system")
list_fonts(family_only = FALSE, source = "google")
}
}
