% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxviolinplot.R
\name{make_long}
\alias{make_long}
\title{Make the wide format data into long format}
\usage{
make_long(data, dodge_by = NULL, dodge_by_sep = "_", facet_by = NULL)
}
\arguments{
\item{data}{A data frame}

\item{dodge_by}{A character string specifying the column name of the data frame to dodge the plot.}

\item{dodge_by_sep}{A character string to concatenate the columns in \code{dodge_by},
if multiple columns are provided and the types of the columns are not logical/(0, 1).}

\item{facet_by}{A character string specifying the column name of the data frame to facet the plot.}
}
\value{
Transformed data with dodge_by and facet_by attributes
}
\description{
Make the wide format data into long format
}
\keyword{internal}
