% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{map.plot.us.size}
\alias{map.plot.us.size}
\title{The US map with fill = size0.}
\usage{
map.plot.us.size(data0 = aeronetdt, size0 = "Ndays", sat0 = sat,
  NEMIA = T)
}
\arguments{
\item{data0}{data}

\item{size0}{variable for point size}

\item{sat0}{name of satellite for labeling}

\item{NEMIA}{default to FALSE plot NEMIA or US}
}
\description{
The US map with fill = size0.
}
