% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_US_NEMIA}
\alias{map_US_NEMIA}
\title{map_US_NEMIA: saved terrain map image from Google Maps}
\format{An object of class \code{ggmap} (inherits from \code{raster}) with 1280 rows and 1280 columns.}
\usage{
map_US_NEMIA
}
\description{
map_US_NEMIA: saved terrain map image from Google Maps
}
\keyword{savedMap}
