% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEMIA_sf_poly}
\alias{NEMIA_sf_poly}
\title{NEMIA_sf_poly.rda: NEMIA sf shape file}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 13 rows and 84 columns.}
\usage{
NEMIA_sf_poly
}
\description{
NEMIA_sf_poly.rda: NEMIA sf shape file
}
\keyword{savedMap}
