% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{prepare.bin}
\alias{prepare.bin}
\title{prepare the folder by day or by station.}
\usage{
prepare.bin(modeldt1, by, k_fold)
}
\arguments{
\item{modeldt1}{dataset}

\item{by}{must be either stn or day}

\item{k_fold}{k fold}
}
\value{
list of a dataset with group assignment and a \code{bin_list} showing what are in each bin
}
\description{
Internal function for cv
}
