% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_keep_na_empty}
\alias{process_keep_na_empty}
\title{process_keep_na_empty
Process keep_na and keep_empty to data}
\usage{
process_keep_na_empty(data, keep_na = NULL, keep_empty = NULL, col = NULL)
}
\arguments{
\item{data}{data frame}

\item{keep_na}{List of keep_na}

\item{keep_empty}{List of keep_empty}
}
\value{
processed data frame
}
\description{
process_keep_na_empty
Process keep_na and keep_empty to data
}
\keyword{internal}
