% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_theme}
\alias{process_theme}
\title{Process theme to allow 'ggplot2::theme_minimal' to work}
\usage{
process_theme(theme)
}
\arguments{
\item{theme}{The theme to process}
}
\value{
The processed theme
}
\description{
Process theme to allow 'ggplot2::theme_minimal' to work
}
\keyword{internal}
