% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{report.r.squared}
\alias{report.r.squared}
\title{report R squared}
\usage{
report.r.squared(dataXY, nround = 2)
}
\arguments{
\item{dataXY}{the dataset, modeldt1_wPred from the model above}

\item{nround}{round the results}
}
\description{
report R squared
}
