% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{rfe.fit}
\alias{rfe.fit}
\title{A wrapped function to run xgboost model.}
\usage{
rfe.fit(X, Y, xgb_param, verbose = FALSE)
}
\arguments{
\item{X}{predictors, Notice that \strong{should NOT contains Y}.}

\item{Y}{dependent variable Y.}

\item{xgb_param}{a list of hyperparameters selected, contains nrounds}
}
\value{
xgboost model object
}
\description{
A wrapped function to run xgboost model.
}
