% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{rfe.rmse.plot}
\alias{rfe.rmse.plot}
\title{Show RFE results: make plots of rmse for new RFE.
updated on (19.04.06). only one line is plotted.}
\usage{
rfe.rmse.plot(rmse_rfe = rferesults$rmse_rfe)
}
\arguments{
\item{rmse_rfe}{use rferesults$rmse_rfe}
}
\value{
ggplot2 object
}
\description{
Show RFE results: make plots of rmse for new RFE.
updated on (19.04.06). only one line is plotted.
}
