% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{run.k.fold.cv}
\alias{run.k.fold.cv}
\title{core cv function what uses \code{xgboost.dart.cvtune} to run cv for one fold.}
\usage{
run.k.fold.cv(sat, k_fold, run_param_cv, dataXY_df, y_var, index_train,
  index_test, xgb_threads, ...)
}
\arguments{
\item{sat}{Name of the satellite, just for labelling purpose, for example "terra"}

\item{k_fold}{default to 5 k_fold cross-validation}

\item{run_param_cv}{run parameter search or not, default to TRUE,
but set to FALSE during the RFE process, could be changed to TRUE in the
code to allow parameter tuning during RFE process. May not be necessary.}

\item{dataXY_df}{dataset with only X and Y}

\item{y_var}{the y variables, for example y_var = "AOD_diff"}

\item{index_train}{index for training}

\item{index_test}{index for testing}

\item{xgb_threads}{xgb_threads passed from parent function}

\item{...}{other arguments}
}
\description{
internal function.
}
