% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{run.k.fold.cv.rfe.wrap}
\alias{run.k.fold.cv.rfe.wrap}
\title{run.k.fold.cv.rfe.wrap: wrap cv (the \code{run.k.fold.cv} function) with rfe.}
\usage{
run.k.fold.cv.rfe.wrap(modeldt1, sat, y_var, features0, k_fold = 5,
  stn_var = NULL, day_var = NULL, run_param_cv = T, run_rfe = T,
  predict_whole_model = T)
}
\arguments{
\item{modeldt1}{the dataset}

\item{sat}{Name of the satellite, just for labelling purpose, for example "terra"}

\item{y_var}{the y variables, for example y_var = "AOD_diff"}

\item{features0}{the features to use}

\item{k_fold}{default to 5 k_fold cross-validation}

\item{stn_var}{the variable presenting stations if cv by stn}

\item{day_var}{the variable presenting dayint if cv by day,
if provide both, will run cv by station. day_var should be date, or int.}

\item{run_param_cv}{run parameter search or not, default to TRUE,
but set to FALSE during the RFE process, could be changed to TRUE in the
code to allow parameter tuning during RFE process. May not be necessary.}

\item{run_rfe}{default to TRUE, run RFE}

\item{predict_whole_model}{default to TRUE, fit the whole model in the end
using selected features from RFE}
}
\value{
a list of object:
\itemize{
\item sat and by: record of satellite and cv by what
\item time_spent: time spent to run the cv
\item bin_list: how was the bins divided
\item modeldt1_wPred: dt of the data with predicted values from cv
\item rmse_all_folds: overall rmse from cv
\item features_rank_rfe: the features ranked by RFE
\item features_rank_rfe_record: what were selected during each round of RFE
\item shap_score: the shap values dataset from cv
\item xgb_param_list1: each folder's hyperparameters in list
\item xgb_param_list2 = unlist(xgb_param_list1)
\item rmse_rfe: rmse from each round of RFE, used to select features
\item var_selected_rfe: features selected using RFE, used to fit the whole model
\item xgb_param_dart_whole_model
\item features_rank_full_model: SHAP rank of features if just run the whole model
not based on RFE
}
}
\description{
Version 3 CV with RFE (recursive feature selection using SHAP),
use random search by \code{xgboost.dart.cvtune}.
}
