% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthis.R
\name{scatter.plot.diagonal}
\alias{scatter.plot.diagonal}
\title{make customized scatter plot with diagonal line and R2 printed.}
\usage{
scatter.plot.diagonal(data, x, y, dilute = F, add_abline = F,
  add_hist = T)
}
\description{
make customized scatter plot with diagonal line and R2 printed.
}
\examples{
scatter.plot.diagonal(data = datam1_terra, x = "col_water_cm", y = "Column_WV", add_abline = T)


}
