% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthis.R
\name{scatter.plot.simple}
\alias{scatter.plot.simple}
\title{simple scatter plot, adding marginal histogram by default.}
\usage{
scatter.plot.simple(data, x, y, size0 = 0.1, alpha0 = 0.3,
  dilute = FALSE, add_hist = TRUE)
}
\arguments{
\item{data}{dataset}

\item{size0}{point size, default to 1 of nobs<1000, 0.4 if nobs>1000.}

\item{dilute}{a number or logical, dafault to TRUE, will plot \code{nrow(data_long)/dilute} data. For example, if dilute = 5 will plot 1/5 of the data.}
}
\description{
simple scatter plot, adding marginal histogram by default.
}
\examples{
scatter.plot.simple(data = datam1_terra, x = "col_water_cm", y = "Column_WV")

}
