% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotthis.R
\name{sf.plot.nyc}
\alias{sf.plot.nyc}
\title{plot only the NYC map}
\usage{
sf.plot.nyc(data, lon = "lon", lat = "lat", zoom_in_option = 2,
  fill = NULL)
}
\description{
plot only the NYC map
}
\examples{
sf.plot.nyc(datam1_terra)

}
