% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjcolors}
\alias{adjcolors}
\title{Convert a color with arbitrary transparency to a fixed color}
\usage{
adjcolors(colors, alpha)
}
\arguments{
\item{colors}{Color vectors.}

\item{alpha}{Alpha level ranging from 0 to 1.}
}
\description{
This function takes a vector of colors and an alpha level and converts the colors
to fixed colors with the specified alpha level.
}
\examples{
colors <- c("red", "blue", "green")
plotthis:::adjcolors(colors, 0.5)
ggplot2::alpha(colors, 0.5)

show_palettes(list(
    "raw" = colors,
    "adjcolors" = plotthis:::adjcolors(colors, 0.5),
    "ggplot2::alpha" = ggplot2::alpha(colors, 0.5)
))

}
\keyword{internal}
