% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustreeplot.R
\name{ClustreePlotAtomic}
\alias{ClustreePlotAtomic}
\title{Atomic function for clustree plot}
\usage{
ClustreePlotAtomic(
  data,
  prefix,
  flip = FALSE,
  alpha = 0.85,
  palette = "Paired",
  palcolor = NULL,
  edge_palette = "Spectral",
  edge_palcolor = NULL,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  theme = "theme_this",
  theme_args = list(),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{prefix}{A character string of the prefix of the columns to plot.
The columns with the prefix will be used to plot the tree.}

\item{flip}{A logical value to flip the tree.}

\item{alpha}{A numeric value of the transparency of the nodes.
Only used when \code{node_alpha} is not provided in \code{...}.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{edge_palette}{A character string of the palette name to color the edges.}

\item{edge_palcolor}{A character vector of colors to color the edges.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_this".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{...}{Other arguments passed to \code{clustree::clustree}.}
}
\description{
Atomic function for clustree plot
}
\keyword{internal}
