% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine plots into one}
\usage{
combine_plots(
  plots,
  combine,
  nrow,
  ncol,
  byrow,
  axes,
  axis_titles,
  guides,
  recalc_size = TRUE
)
}
\arguments{
\item{plots}{A list of plots}

\item{combine}{Whether to combine the plots into one}

\item{nrow}{The number of rows in the combined plot}

\item{ncol}{The number of columns in the combined plot}

\item{byrow}{Whether to fill the plots by row}

\item{recalc_size}{Whether to re-calculate the size of the combined plot}
}
\value{
The faceted plot. If guess_size is TRUE, attr(p, "height") and attr(p, "width") will be set
}
\description{
Combine plots into one
}
\keyword{internal}
