% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{DensityHistoPlotAtomic}
\alias{DensityHistoPlotAtomic}
\title{Atomic density/histogram plot}
\usage{
DensityHistoPlotAtomic(
  data,
  x,
  group_by = NULL,
  group_by_sep = "_",
  group_name = NULL,
  type = c("density", "histogram"),
  bins = NULL,
  binwidth = NULL,
  flip = FALSE,
  add_lines = TRUE,
  line_height = 0.025,
  line_alpha = 1,
  line_width = 0.1,
  palette = "Paired",
  palcolor = NULL,
  alpha = 0.5,
  theme = "theme_this",
  theme_args = list(),
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  expand = c(bottom = 0, left = 0, right = 0),
  facet_by = NULL,
  facet_scales = "fixed",
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_byrow = TRUE,
  legend.position = ifelse(is.null(group_by), "none", "right"),
  legend.direction = "vertical",
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A character string specifying the column name for the values
A numeric column is expected.}

\item{group_by}{A character string specifying the column name to group the data}

\item{group_by_sep}{A character string to concatenate the columns in \code{group_by} if multiple columns are provided}

\item{group_name}{A character string to name the legend of group_by}

\item{type}{A character string specifying the type of plot. Default is "density".
Other options are "histogram".}

\item{bins}{A numeric value specifying the number of bins for the histogram.}

\item{binwidth}{A numeric value specifying the width of the bins for the histogram.}

\item{flip}{A logical value. If TRUE, the plot will be flipped.}

\item{add_lines}{A logical value. If TRUE, add lines to the plot to show the data distribution on the bottom.}

\item{line_height}{A numeric value specifying the height of the lines. The actual height will be calculated based on the maximum density or count.}

\item{line_alpha}{A numeric value specifying the alpha of the lines.}

\item{line_width}{A numeric value specifying the width of the lines.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{alpha}{A numeric value specifying the transparency of the plot.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_this".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{expand}{The values to expand the x and y axes. It is like CSS padding.
When a single value is provided, it is used for both axes on both sides.
When two values are provided, the first value is used for the top/bottom side and the second value is used for the left/right side.
When three values are provided, the first value is used for the top side, the second value is used for the left/right side, and the third value is used for the bottom side.
When four values are provided, the values are used for the top, right, bottom, and left sides, respectively.
You can also use a named vector to specify the values for each side.
When the axis is discrete, the values will be applied as 'add' to the 'expansion' function.
When the axis is continuous, the values will be applied as 'mult' to the 'expansion' function.
See also \url{https://ggplot2.tidyverse.org/reference/expansion.html}}

\item{facet_by}{A character string specifying the column name of the data frame to facet the plot.
Otherwise, the data will be split by \code{split_by} and generate multiple plots
and combine them into one using \code{patchwork::wrap_plots}}

\item{facet_scales}{Whether to scale the axes of facets. Default is "fixed"
Other options are "free", "free_x", "free_y". See \code{ggplot2::facet_wrap}}

\item{facet_ncol}{A numeric value specifying the number of columns in the facet.
When facet_by is a single column and facet_wrap is used.}

\item{facet_nrow}{A numeric value specifying the number of rows in the facet.
When facet_by is a single column and facet_wrap is used.}

\item{facet_byrow}{A logical value indicating whether to fill the plots by row. Default is TRUE.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{...}{Additional arguments.}
}
\description{
Atomic density/histogram plot
}
\keyword{internal}
