% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venndiagram.R
\name{detect_venn_datatype}
\alias{detect_venn_datatype}
\title{Detect the datatype of the input data of Venn diagram}
\usage{
detect_venn_datatype(data, group_by = NULL, id_by = NULL)
}
\arguments{
\item{data}{A data frame or a list}

\item{group_by}{A character string specifying the column name of the data frame to group the data.}

\item{id_by}{A character string specifying the column name of the data frame to identify the instances.
Required when \code{group_by} is a single column and data is a data frame.}
}
\value{
A character string indicating the datatype of the input data or error message if invalid.
Possible values are "long", "wide", "list"
"long" indicates the data is in long format.
"wide" indicates the data is in wide format.
"list" indicates the data is a list.
}
\description{
Detect the datatype of the input data of Venn diagram
}
\keyword{internal}
