% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{.prep_annotations}
\alias{.prep_annotations}
\title{Prepare and normalize annotation arguments}
\usage{
.prep_annotations(
  which = c("row", "column"),
  annotation = NULL,
  annotation_side = NULL,
  annotation_palette = NULL,
  annotation_palcolor = NULL,
  annotation_type = NULL,
  annotation_params = NULL,
  annotation_agg = NULL,
  row_key = NULL,
  rsplit_key = NULL,
  col_key = NULL,
  csplit_key = NULL,
  data = NULL
)
}
\arguments{
\item{which}{\code{"row"} or \code{"column"}.}

\item{annotation}{The new structured annotation list.}

\item{annotation_side}{Deprecated. Old side argument.}

\item{annotation_palette}{Deprecated. Old palette argument.}

\item{annotation_palcolor}{Deprecated. Old palcolor argument.}

\item{annotation_type}{Deprecated. Old type argument.}

\item{annotation_params}{Deprecated. Old params argument.}

\item{annotation_agg}{Deprecated. Old agg argument.}

\item{row_key}{The actual column name for \code{rows_by}.}

\item{rsplit_key}{The actual column name for \code{rows_split_by}.}

\item{col_key}{The actual column name for \code{columns_by}.}

\item{csplit_key}{The actual column name for \code{columns_split_by}.}

\item{data}{The data frame used for column validation.}
}
\value{
A list with components \code{annotation}, \code{annotation_type},
\code{annotation_side}, \code{annotation_palette}, \code{annotation_palcolor},
\code{annotation_agg}, \code{annotation_params}, and \code{enabled}.
}
\description{
Consolidates the new structured annotation list format with backward-compatible
deprecation support for old flat arguments. Handles \code{TRUE}/\code{FALSE}
shortcuts, \code{.default} inheritance (with recursive \code{params} merge),
alias resolution, and shortcut expansion.
}
\keyword{internal}
