% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfe.R
\name{helper.pack.bins}
\alias{helper.pack.bins}
\title{assign groups numbers to each station by their number of observations.}
\usage{
helper.pack.bins(object.sizes, n.bins)
}
\arguments{
\item{object.sizes}{a vector of object size}

\item{n.bins}{number of bins (e.g. 5 folders )}
}
\value{
a vector of group assignment
}
\description{
Internal function for cv
}
\examples{
helper.pack.bins(seq(1:10), n.bins = 4)
}
