% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{Scatter Plot}
\usage{
ScatterPlot(
  data,
  x,
  y,
  flip = FALSE,
  split_by = NULL,
  split_by_sep = "_",
  size_name = NULL,
  fill_name = NULL,
  fill_cutoff_name = NULL,
  size_by = NULL,
  fill_by = NULL,
  fill_cutoff = NULL,
  fill_reverse = FALSE,
  theme = "theme_this",
  theme_args = list(),
  palette = "Spectral",
  palcolor = NULL,
  alpha = 1,
  facet_by = NULL,
  facet_scales = "fixed",
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_byrow = TRUE,
  x_text_angle = 0,
  seed = 8525,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  keep_empty = FALSE,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A character vector specifying the column to use for the x-axis.
Must be a numeric column.}

\item{y}{A character vector specifying the column to use for the y-axis.
Must be a numeric column.}

\item{flip}{A logical value indicating whether to flip the x and y axes. Default is FALSE.}

\item{split_by}{The column(s) to split data by and plot separately.}

\item{split_by_sep}{The separator for multiple split_by columns. See \code{\link{split_by}}}

\item{size_name}{A character vector specifying the name for the size legend.}

\item{fill_name}{A character vector specifying the name for the fill legend.}

\item{fill_cutoff_name}{A character vector specifying the name for the fill cutoff legend.}

\item{size_by}{Which column to use as the size of the dots. It must be a numeric column.
If not provided, the size will be the count of the instances for each 'y' in 'x'.}

\item{fill_by}{Which column to use as the fill the dots. It must be a numeric column.
If not provided, all dots will be filled with the same color at the middle of the palette.}

\item{fill_cutoff}{A numeric value specifying the cutoff for the fill column.}

\item{fill_reverse}{A logical value indicating whether to reverse the fill direction. Default is FALSE.
By default, the fill direction is "up". If TRUE, the fill direction is "down".
When the direction is "up", the values less than the cutoff will be filled with grey.
When the direction is "down", the values greater than the cutoff will be filled with grey.}

\item{theme}{A character string or a theme class (i.e. ggplot2::theme_classic) specifying the theme to use.
Default is "theme_this".}

\item{theme_args}{A list of arguments to pass to the theme function.}

\item{palette}{A character string specifying the palette to use.}

\item{palcolor}{A character string specifying the color to use in the palette.}

\item{alpha}{A numeric value specifying the transparency of the plot.}

\item{facet_by}{A character string specifying the column name of the data frame to facet the plot.
Otherwise, the data will be split by \code{\link{split_by}} and generate multiple plots
and combine them into one using \code{patchwork::wrap_plots}}

\item{facet_scales}{Whether to scale the axes of facets. Default is "fixed"
Other options are "free", "free_x", "free_y". See \code{\link{ggplot2::facet_wrap}}}

\item{facet_ncol}{A numeric value specifying the number of columns in the facet.
When facet_by is a single column and facet_wrap is used.}

\item{facet_nrow}{A numeric value specifying the number of rows in the facet.
When facet_by is a single column and facet_wrap is used.}

\item{facet_byrow}{A logical value indicating whether to fill the plots by row. Default is TRUE.}

\item{x_text_angle}{A numeric value specifying the angle of the x-axis text.}

\item{seed}{The random seed to use. Default is 8525.}

\item{aspect.ratio}{A numeric value specifying the aspect ratio of the plot.}

\item{legend.position}{A character string specifying the position of the legend.}

\item{legend.direction}{A character string specifying the direction of the legend.}

\item{title}{A character string specifying the title of the plot.}

\item{subtitle}{A character string specifying the subtitle of the plot.}

\item{xlab}{A character string specifying the x-axis label.}

\item{ylab}{A character string specifying the y-axis label.}

\item{keep_empty}{A logical value indicating whether to keep empty groups.
If FALSE, empty groups will be removed.}

\item{combine}{Whether to combine the plots into one when facet is FALSE. Default is TRUE.}

\item{nrow}{A numeric value specifying the number of rows in the facet.}

\item{ncol}{A numeric value specifying the number of columns in the facet.}

\item{byrow}{A logical value indicating whether to fill the plots by row.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object or wrap_plots object or a list of ggplot objects
}
\description{
A alias of DotPlot when both x and y are numeric.
}
\examples{
mtcars <- datasets::mtcars
ScatterPlot(mtcars, x = "qsec", y = "drat", size_by = "wt",
            fill_by = "mpg")
ScatterPlot(mtcars, x = "qsec", y = "drat", size_by = "wt",
            fill_by = "mpg", fill_cutoff = 18)
}
