% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaps.R
\name{sf.plot.nyc}
\alias{sf.plot.nyc}
\title{plot only the NYC map}
\usage{
sf.plot.nyc(data, lon = "lon", lat = "lat", zoom_in_option = 2,
  fill = NULL)
}
\arguments{
\item{data}{data}

\item{lon}{lon}

\item{lat}{lat}

\item{zoom_in_option}{offer 1,2,3 different scope of the city}

\item{fill}{optional to fill by variable, default to NULL, plot points}
}
\description{
plot only the NYC map
}
\examples{
sf.plot.nyc(datam1_terra)
sf.plot.nyc(datam1_terra, fill = "elev")

}
